Agni Flight Computer V2.
This flight computer is designed specifically for GUI purposes, and it is my second PCB board. Since it's an early version, it's not fully SMD—I have used a mix of SMD and through-hole (TH) components.

In the next version, Agni Flight Computer V2.1, the design will be 50% SMD and 50% TH, as I continue improving my PCB design skills. Through this process, I am learning how to design PCBs using EasyEDA and KiCad. 🚀